Comply with Java Coding Standards
Make the code look like it was written by a single person by complying with Java coding standards.
