+++ src/main/java/customexceptions/WrongCommandException.java
@@ -1,4 +1,4 @@
-package CustomExceptions;
+package customexceptions;
 public class WrongCommandException extends Exception {
     public WrongCommandException(String command) {
         super(command + " is not a proper command! " +
