+++ src/main/java/parser/Parser.java
@@ -1,4 +1,4 @@
-package Parser;
+package parser;
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
