+++ src/main/java/storage/Storage.java
@@ -1,11 +1,11 @@
-package Storage;
+package storage;
 import java.io.IOException;
 import java.util.Scanner;
 import java.io.File;
 import java.io.FileWriter;
-import Tasks.*;
-import Parser.Parser;
-import Ui.Ui;
+import tasks.*;
+import parser.Parser;
+import ui.Ui;
 
 public class Storage {
     private FileWriter pw;
