+++ src/main/java/tasks/Task.java
@@ -1,4 +1,4 @@
-package Tasks;
+package tasks;
 public class Task {
     private boolean isDone;
     private final String NAME;
