+++ src/test/java/parser/ParserTest.java
@@ -1,4 +1,4 @@
-package Parser;
+package parser;
 
 import org.junit.jupiter.api.Test;
 import static org.junit.jupiter.api.Assertions.*;
